+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 * **Quality of Life:**
   - Player view no longer abruptly jolts when running across very shallow floor height changes
 
+* **Rendering:**
+  - Improved palettization of PNG graphics.
+
 ## Bug Fixes
 
 * Fixed setup desktop action exiting immediately when run from the AppImage